Add diagnostic pragma for unused variable
When compiling with -Wall and -Wextra GCC flags for ESP32 targets, the "initialized" variable as unused even though it is used in several macros. Adding this diagnostic pragma suppress the warning. A better solution would be to re-architect the library to avoid the WAIT_FOR_SLAVE_SELECT macro in favor of a proper function but it is unclear how this would impact other targets.